Replace rule `condition-wrapping` with `expression-operand-wrapping`
Logical and arithmetic expressions should wrap their operands identically. Also, the wrapping should not be limited to conditions but to any expression

Closes #3000
